Update README.md (#25)
I think that people are going to skim this looking for relevant information.  They're going to jump straight to either the backbone or react section which previously would have led them to implement the incorrect webpack way.  I also think versioning and customization should be pushed out of the way because they are less important to everyone.